src/main/java/com/crud/tasks/service/DbService.java
package com.crud.tasks.service;


import com.crud.tasks.domain.Task;
import com.crud.tasks.exception.TaskNotFoundException;
import com.crud.tasks.repository.TaskRepository;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

import java.util.List;

@Service
public class DbService {
    @Autowired
    private TaskRepository repository;

    public List<Task> getAllTasks() {
        return repository.findAll();
    }

    public Task getTask(final Long id) throws TaskNotFoundException {
        return repository.findById(id).orElseThrow(TaskNotFoundException::new);
    }

    public Task saveTask(final Task task) {
        return repository.save(task);
    }

    public void deleteById(final Long id) {
        repository.deleteById(id);
    }

    public Task updateTask(Task task) throws TaskNotFoundException {
        getTask(task.getId());
        return saveTask(task);
    }



}
